--- knora/000__init__.py
@@ -1,6 +1,0 @@
-#from .knora import KnoraStandoffXml
-#from .knora import Knora
-#from .knora import Sipi
-#from .knora import BulkImport
-#from .knora import IrisLookup
-#from .knora import ListsLookup
