+++ knora/libknora/knoraConsoleModules/BUILD
