+++ knora/libknora/knoraConsoleModules/ResourcePanel.py
@@ -6,18 +6,18 @@ import wx
 from pprint import pprint
 import copy
 
-from knora.models.helpers import Actions, BaseError, Context, Cardinality, LastModificationDate
-from knora.models.langstring import Languages, LangStringParam, LangString
-from knora.models.connection import Connection, Error
-from knora.models.project import Project
-from knora.models.listnode import ListNode
-from knora.models.group import Group
-from knora.models.user import User
-from knora.models.ontology import Ontology
-from knora.models.propertyclass import PropertyClass
-from knora.models.resourceclass import ResourceClass, HasProperty
+from ..models.helpers import Actions, BaseError, Context, Cardinality, LastModificationDate
+from ..models.langstring import Languages, LangStringParam, LangString
+from ..models.connection import Connection, Error
+from ..models.project import Project
+from ..models.listnode import ListNode
+from ..models.group import Group
+from ..models.user import User
+from ..models.ontology import Ontology
+from ..models.propertyclass import PropertyClass
+from ..models.resourceclass import ResourceClass, HasProperty
 
-from knora.knoraConsoleModules.KnDialogControl import show_error, KnDialogControl, KnDialogTextCtrl, KnDialogChoice, \
+from ..knoraConsoleModules.KnDialogControl import show_error, KnDialogControl, KnDialogTextCtrl, KnDialogChoice, \
     KnDialogChoiceArr, KnDialogCheckBox, KnCollapsiblePicker, KnDialogStaticText, KnDialogSuperResourceClasses, \
     KnDialogGuiAttributes, KnDialogLangStringCtrl, KnDialogHasProperty, PropertyStatus
