+++ knora/libknora/models/listnode.py
@@ -8,10 +8,10 @@ from enum import Enum, unique
 from urllib.parse import quote_plus
 from pprint import pprint
 
-from knora.models.helpers import Actions, BaseError
-from knora.models.langstring import Languages, LangStringParam, LangString
-from knora.models.connection import Connection
-from knora.models.project import Project
+from ..models.helpers import Actions, BaseError
+from ..models.langstring import Languages, LangStringParam, LangString
+from ..models.connection import Connection
+from ..models.project import Project
 
 class SetEncoder(json.JSONEncoder):
     def default(self, obj):
