+++ knora/libknora/utils/knora-schema-lists.json
