+++ knora/libknora/utils/knora-schema.json
