+++ knora/libknora/utils/onto_commons.py
@@ -1,22 +1,8 @@
-import os
 from typing import List, Set, Dict, Tuple, Optional
-from pprint import pprint
-import argparse
-import json
-from jsonschema import validate
-import sys
-import re
 
-from knora.models.helpers import Actions, BaseError, Context, Cardinality
-from knora.models.langstring import Languages, LangStringParam, LangString
-from knora.models.connection import Connection, Error
-from knora.models.project import Project
-from knora.models.listnode import ListNode
-from knora.models.group import Group
-from knora.models.user import User
-from knora.models.ontology import Ontology
-from knora.models.propertyclass import PropertyClass
-from knora.models.resourceclass import ResourceClass
+from ..models.connection import Connection
+from ..models.project import Project
+from ..models.listnode import ListNode
 
 def list_creator(con: Connection, project: Project, parent_node: ListNode, nodes: List[dict]):
     nodelist = []
