+++ knora/libknora/utils/onto_validate.py
@@ -1,22 +1,6 @@
 import os
-from typing import List, Set, Dict, Tuple, Optional
-from pprint import pprint
-import argparse
 import json
 from jsonschema import validate
-import sys
-import re
-
-from knora.models.helpers import Actions, BaseError, Context, Cardinality
-from knora.models.langstring import Languages, LangStringParam, LangString
-from knora.models.connection import Connection, Error
-from knora.models.project import Project
-from knora.models.listnode import ListNode
-from knora.models.group import Group
-from knora.models.user import User
-from knora.models.ontology import Ontology
-from knora.models.propertyclass import PropertyClass
-from knora.models.resourceclass import ResourceClass
 
 
 def validate_list(input_file: str) -> None:
